Revert "Update to Linux kernel 2.6.34.x.. first wave"
This reverts commit b28dcbd767cc48df44d9d0700e554632637181c5.

+++ arch/microblaze/kernel/cpu/pvr.c
@@ -45,7 +45,7 @@
 
 int cpu_has_pvr(void)
 {
-	unsigned long flags;
+	unsigned flags;
 	unsigned pvr0;
 
 	local_save_flags(flags);
